refactor: add `at.rs`
